backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
torch
transformers
pdfplumber
reportlab
python-docx
gTTS
pytesseract
Pillow